% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Method1}
\alias{Method1}
\title{Method1 function}
\usage{
Method1(MyData, AgeRange, Plot = F, Color = MyData$Param$Color)
}
\arguments{
\item{MyData}{The list returned by the AddReference() function.}

\item{AgeRange}{Age range used for the calculation of the SMR.}

\item{Plot}{If set to TRUE, a sub-directory will be created in the working directory. This sub-directory will contains png plots corresponding to the smoothed surface.}

\item{Color}{The color that will be used for the plots (HTML notation).}
}
\description{
This function fits the Qxt using method 1 (SMR method, see reference).
}

