% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S06.R
\name{Method4A}
\alias{Method4A}
\title{Method4A function}
\usage{
Method4A(MyData, AgeRange, AgeCrit, ShowPlot = F)
}
\arguments{
\item{MyData}{The list returned by the AddReference() function.}

\item{AgeRange}{Age range used for the construction of the life table.}

\item{AgeCrit}{Age range for the comparison of adjusted mortality and observed mortality.}

\item{ShowPlot}{AIC plots and plots allowing to judge about the fit.}
}
\description{
This function fits the Qxt using method 4 (first step) (non-parametric smoothing, see reference).
}

