% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Method4B}
\alias{Method4B}
\title{Method4B function}
\usage{
Method4B(PartOne, MyData, OptMale, OptFemale, Plot = F, ShowPlot = F,
  Color = MyData$Param$Color)
}
\arguments{
\item{PartOne}{The list returned by the Method4A() function.}

\item{MyData}{The list returned by the AddReference() function.}

\item{OptMale}{Optimal smoothing parameters, obtained from the graphics generated by Method4A() for the male population.}

\item{OptFemale}{Optimal smoothing parameters, obtained from the graphics generated by Method4A() for the female population.}

\item{Plot}{If set to TRUE, a sub-directory will be created in the working directory. This sub-directory will contains png plots corresponding to the smoothed surface.}

\item{ShowPlot}{If true, show plots.}

\item{Color}{The color that will be used for the plots (HTML notation).}
}
\description{
This function fits the Qxt using method 4 (second step) (non-parametric smoothing, see reference).
}

