% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S01.R
\name{NoCompletion}
\alias{NoCompletion}
\title{NoCompletion function}
\usage{
NoCompletion(OutputMethod, MyData, Color = MyData$Param$Color, Plot = F,
  Excel = F)
}
\arguments{
\item{OutputMethod}{The list returned by one of these functions : Method1(), Method2(), Method3() or Method4B().}

\item{MyData}{The list returned by the AddReference() function.}

\item{Color}{The color that will be used for the plots (HTML notation).}

\item{Plot}{If TRUE, final mortality surfaces will be saved in Results/Graphics/FinalTables}

\item{Excel}{If TRUE, final tables will be saved in Results/Excel/FinalTables.xlsx}
}
\description{
This function allows to keep the adjustment used by the locating method for high ages (for methods 1, 2 or 3).
}

