\name{ReferenceFemale}
\alias{ReferenceFemale}
\docType{data}
\title{ReferenceFemale used for the exemple.}
\description{This data corresponds to an adjusted version of the French national demographic projections INSEE 2060 for the female population.}
\usage{data(ReferenceFemale)}
\examples{data(ReferenceFemale)}
\keyword{datasets}
