\name{ReferenceMale}
\alias{ReferenceMale}
\docType{data}
\title{ReferenceMale used for the exemple.}
\description{This data corresponds to an adjusted version of the French national demographic projections INSEE 2060 for the male population.}
\usage{data(ReferenceMale)}
\examples{data(ReferenceMale)}
\keyword{datasets}
