% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ValidationLevel2}
\alias{ValidationLevel2}
\title{ValidationLevel2 function}
\usage{
ValidationLevel2(OutputMethod, MyData, ValCrit, AgeCrit, Excel = F)
}
\arguments{
\item{OutputMethod}{The list returned by one of these functions : Method1(), Method2(), Method3() or Method4B().}

\item{MyData}{The list returned by the AddReference() function.}

\item{ValCrit}{Critical value for the comparison of adjusted mortality and observed mortality.}

\item{AgeCrit}{Age range for the comparison of adjusted mortality and observed mortality.}

\item{Excel}{If set to TRUE, a sub-directory will be created in the working directory. This sub-directory will contains excel files describing the validation analysis.}
}
\description{
This function performs the second level of validation on the returned value of one of these functions : Method1(), Method2(), Method3() or Method4B() (see reference).
}

