% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S10.R
\name{ValidationLevel3}
\alias{ValidationLevel3}
\title{ValidationLevel3 function}
\usage{
ValidationLevel3(FinalMethod, MyData, Plot = F, Color = MyData$Param$Color,
  Excel = F)
}
\arguments{
\item{FinalMethod}{The list returned by the CompletionB() function.}

\item{MyData}{The list returned by the AddReference() function.}

\item{Plot}{If set to TRUE, a sub-directory will be created in the working directory. This sub-directory will contains png plots describing the validation analysis.}

\item{Color}{The color that will be used for the plots (HTML notation).}

\item{Excel}{If set to TRUE, a sub-directory will be created in the working directory. This sub-directory will contains excel files describing the validation analysis.}
}
\description{
This function performs the third level of validation on the returned value of one of these functions : Method1(), Method2(), Method3() or Method4B().
}

