% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S02.R
\name{AddReference}
\alias{AddReference}
\title{AddReference function.}
\usage{
AddReference(History, ReferenceMale = NULL, ReferenceFemale = NULL)
}
\arguments{
\item{History}{History as returned by the ReadHistory function.}

\item{ReferenceMale}{data.frame representing the reference table. See data(ReferenceMale) for the format.}

\item{ReferenceFemale}{data.frame representing the reference table. See data(ReferenceFemale) for the format.}
}
\description{
This function imports reference tables.
}

