% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/S11.R
\name{Dispersion}
\alias{Dispersion}
\title{Dispersion function}
\usage{
Dispersion(FinalMethod, MyData, NbSim, CompletionTable = T, Plot = F,
  Color = MyData$Param$Color)
}
\arguments{
\item{FinalMethod}{The list returned by the CompletionB() function.}

\item{MyData}{The list returned by the AddReference() function.}

\item{NbSim}{The number of simulations for the Dispersion.}

\item{CompletionTable}{If TRUE, apply completion}

\item{Plot}{If set to TRUE, a sub-directory will be created in the working directory. This sub-directory will contains png plots describing the validation analysis.}

\item{Color}{The color that will be used for the plots (HTML notation).}
}
\description{
This function allows to calculate confidence intervals for period life expectancies.
}

