\name{SurfacePlot}
\alias{SurfacePlot}
\title{SurfacePlot function}
\usage{
  SurfacePlot(xx, zexpr, mainexpr, axis, cc)
}
\arguments{
  \item{xx}{data as matrix.}

  \item{zexpr}{Title of z axis.}

  \item{mainexpr}{Name for the graphic.}

  \item{axis}{c(min(abscissa), max(abscissa),
  min(ordinate), max(ordinate)).}

  \item{cc}{Color.}
}
\description{
  Allows to plot a surface.
}

